Added SHMS chambers  --  not correct!!
Added the geometry and variables of SHMS chambers. I believe it is
not yet OK. I just spliced the geometry on the HMS chambers, which
have a different naming scheme because of different wire rotations.
But since the logic is kinda hardcoded, I need to debug it to make
it work for SHMS geometry. It displays wires that were hit, but
I dont think the numbers are OK yet.